+++ src/Controller/CardController.php
@@ -12,23 +12,30 @@ use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
+use App\Repository\CardRepository;
+use App\Repository\CardListRepository;
 
 class CardController extends AbstractController
 {
     /**
-     * @Route("/column/{id}/card", name="card_new", methods={"POST"})
+     * @Route("board/{board_id}/column/{column_id}/card", name="card_new", methods={"POST"})
+     * 
+     * @ParamConverter("board", options={"mapping": {"board_id": "id"}})
+     * @ParamConverter("cardList", options={"mapping": {"column_id": "id"}})
      */
-    public function new(CardList $cardList, Request $request, ValidatorInterface $validator, SerializerInterface $serializer): JsonResponse
+    public function new(Board $board, CardList $cardList, Request $request, ValidatorInterface $validator, SerializerInterface $serializer): JsonResponse
     {
+        $this->denyAccessUnlessGranted('edit', $board);
+
         $parametersAsArray = [];
         if ($content = $request->getContent()) {
             $parametersAsArray = json_decode($content, true);
         }
 
         $card = new Card();
-        $card->setTitle($parametersAsArray['title']);
-        $card->setPosition($parametersAsArray['position']);
-        $card->setCardList($cardList);
+        $card->setTitle($parametersAsArray['title'])
+            ->setPosition($parametersAsArray['position'])
+            ->setCardList($cardList);
 
         $errors = $validator->validate($card);
         if (count($errors) > 0) {
@@ -45,20 +52,23 @@ class CardController extends AbstractController
     }
 
     /**
-     * @Route("/column/{column_id}/card/{card_id}", name="card_edit", methods={"PATCH"})
+     * @Route("board/{board_id}/column/{column_id}/card/{card_id}", name="card_edit", methods={"PATCH"})
      * 
+     * @ParamConverter("board", options={"mapping": {"board_id": "id"}})
      * @ParamConverter("cardList", options={"mapping": {"column_id": "id"}})
      * @ParamConverter("card", options={"mapping": {"card_id": "id"}})
      */
-    public function edit(CardList $cardList, Card $card, Request $request, ValidatorInterface $validator): JsonResponse
+    public function edit(Board $board, CardList $cardList, Card $card, Request $request, ValidatorInterface $validator): JsonResponse
     {
+        $this->denyAccessUnlessGranted('edit', $board);
+
         $parametersAsArray = [];
         if ($content = $request->getContent()) {
             $parametersAsArray = json_decode($content, true);
         }
 
-        $card->setTitle($parametersAsArray['title']);
-        $card->setDescription($parametersAsArray['description']);
+        $card->setTitle($parametersAsArray['title'])
+            ->setDescription($parametersAsArray['description']);
 
         $errors = $validator->validate($card);
         if (count($errors) > 0) {
@@ -73,8 +83,10 @@ class CardController extends AbstractController
     /**
      * @Route("/board/{id}/cards", name="card_edit_batch", methods={"PATCH"})
      */
-    public function editBatch(Board $board, Request $request, ValidatorInterface $validator): JsonResponse
+    public function editBatch(Board $board, CardRepository $cardRepository, CardListRepository $cardListRepository, Request $request, ValidatorInterface $validator): JsonResponse
     {
+        $this->denyAccessUnlessGranted('edit', $board);
+
         $parametersAsArray = [];
         if ($content = $request->getContent()) {
             $parametersAsArray = json_decode($content, true);
@@ -83,11 +95,11 @@ class CardController extends AbstractController
         $entityManager = $this->getDoctrine()->getManager();
 
         foreach ($parametersAsArray as $cardData) {
-            $card = $this->getDoctrine()->getRepository(Card::class)->find($cardData['id']);
+            $card = $cardRepository->find($cardData['id']);
             $card->setPosition($cardData['position']);
 
             if (isset($cardData['card_list_id'])) {
-                $card->setCardList($this->getDoctrine()->getRepository(CardList::class)->find( $cardData['card_list_id']));
+                $card->setCardList($cardListRepository->find( $cardData['card_list_id']));
             }
 
             $errors = $validator->validate($card);
@@ -104,13 +116,16 @@ class CardController extends AbstractController
     }
 
     /**
-     * @Route("/column/{column_id}/card/{card_id}", name="card_delete", methods={"DELETE"})
+     * @Route("board/{board_id}/column/{column_id}/card/{card_id}", name="card_delete", methods={"DELETE"})
      * 
+     * @ParamConverter("board", options={"mapping": {"board_id": "id"}})
      * @ParamConverter("cardList", options={"mapping": {"column_id": "id"}})
      * @ParamConverter("card", options={"mapping": {"card_id": "id"}})
      */
-    public function delete(CardList $cardList, Card $card): JsonResponse
+    public function delete(Board $board, CardList $cardList, Card $card): JsonResponse
     {
+        $this->denyAccessUnlessGranted('edit', $board);
+
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($card);
         $entityManager->flush();
